Fix block quote marker seen as adornment
This fixes an edge case where a blockquote marker on a line just after a
paragraph is seen as an adornment and thus lead to a section being seen.

I found a couple of case in a few library with this.

I think a better fix would be to look back at the title length (is it
possible ? ), or maybe even to require titles to be at least 3 in length
?